upgrade generatePhraseStyle.js to a handle and modularize changeVideo
