+++ src/modules/commit-info.js
@@ -1,1 +1,1 @@
-export const COMMIT_ID = '8be545f53e5d'
+export const COMMIT_ID = '94271ec914bd'
